* move interactive modules from common/common.ts to common/interactions.ts for a better directory tree management
* move window messages corresponding to apis to common/api.ts
